\docType{package}
\name{MICsplines-package}
\alias{MICsplines-package}
\title{
\packageTitle{MICsplines}
}
\description{
The package provides C implementation for the computing of monotonic spline bases, including M-splines, I-splines, and C-splines, denoted by MIC splines. The definitions of the spline bases are described in Meyer (2008). The package also provides the computing of constrained least-squares estimates when a subset of or all of the regression coefficients are constrained to be non-negative, as described in Fraser and Massam (1989).
}

\references{
Fraser, D. A. S. and H. Massam (1989). A mixed primal-dual bases algorithm for regression
under inequality constraints. Application to concave regression. \emph{Scandinavian Journal
of Statistics} 16, 65-74.

Meyer, M. C. (2008). Inference using shape-restricted regression splines. \emph{The Annals of
Applied Statistics} 2, 1013-1033.

}

\keyword{package}


